Re-organized and arranged the render and _serialize methods in order to fix a bug that occured when _serialize is set to true.
When _serialize was set to true in order to automatically serialize all viewVars the resulting ajax response was always {"error":null,"content":"","1":null} where _serialize was set to 1 with a null value. I re-organized the code to what made a little more sense to myself as well as resolved the error.